Replace cacheDir with the cache directory in USB
