+++ cmds/webboot/utils.go
@@ -3,11 +3,16 @@ package main
 import (
 	"fmt"
 	"io"
+	"io/ioutil"
 	"log"
 	"net/http"
 	"net/url"
 	"os"
+	"path/filepath"
 	"strings"
+
+	"github.com/u-root/u-root/pkg/mount"
+	"github.com/u-root/u-root/pkg/mount/block"
 )
 
 // WriteCounter counts the number of bytes written to it. It implements to the io.Writer
@@ -65,4 +70,29 @@ func download(URL, fPath string) error {
 	fmt.Println("\nDone!")
 	verbose("%q is downloaded at %q\n", URL, fPath)
 	return nil
+}
+
+// getCachedDirectory recognizes the usb stick that contains the cached directory from block devices,
+// and return it's mount point.
+func getCachedDirectory() (*mount.MountPoint, error) {
+	blockDevs, err := block.GetBlockDevices()
+	if err != nil {
+		log.Fatal("No available block devices to boot from")
+	}
+
+	mountPoints, err := ioutil.TempDir("", "temp-device-")
+	if err != nil {
+		return nil, fmt.Errorf("cannot create tmpdir: %v", err)
+	}
+
+	for _, device := range blockDevs {
+		mp, err := mount.TryMount("/dev/"+device.Name, filepath.Join(mountPoints, device.Name), "", mount.ReadOnly)
+		if err != nil {
+			continue
+		}
+		if _, err = os.Stat(filepath.Join(mp.Path, "Image")); err == nil {
+			return mp, nil
+		}
+	}
+	return nil, nil
 }
